Final refactoring - Deliver seat
