+++ src/fr/vergne/qualitemeritis/OfferBestSeatsInPriceRangeTest.java
@@ -358,24 +358,6 @@ class OfferBestSeatsInPriceRangeTest {
 		assertEquals(new HashSet<>(expected), new HashSet<>(actual));
 	}
 
-	static class Seat {
-
-		private final Price price;
-
-		public Seat(Price price) {
-			this.price = price;
-		}
-
-		public Price price() {
-			return price;
-		}
-
-		@Override
-		public String toString() {
-			return "Seat[" + hashCode() % 1000 + "]@" + price;
-		}
-	}
-
 	static class SuggestionSystem {
 
 		private final Collection<Seat> seats;
